Use current working directory to get version label
The `get_version_label` function was lifted and shifted from the
applications however, the relative way of finding the version_label file
is no longer valid. We also cannot garuntee where the dmutils package
will be installed so a path relative to this package can't work
consistently.

I have chosen to go with relative to the current working
directory which should work for our use case. Other options are relative
to the calling module file (may not be the same across apps) or to have
the root path passed in (requires a change in each app).